% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunBootBand.R
\docType{data}
\name{data}
\alias{data}
\title{FunBootBand}
\format{
An object of class \code{data.frame} with 11110 rows and 4 columns.
}
\usage{
band(data, k.coef = 50, B = 10, type = "prediction",
cp.begin = 0, alpha = 0.05, iid = TRUE)
}
\arguments{
\item{data}{A data set consisting of n curves of length t. Needs to be a
numerical matrix of dimensions [t, n]. E.g., data[1:101, 1:50] represents 50 curves of lenght 101 points.}

\item{type}{Band type (c("confidence", "prediction", "tolerance"))}

\item{B}{Number of bootstrap iterations (e.g., B = 1000)}

\item{iid}{Assume independent and identically distributed (iid) curves or not
(iid = c(TRUE, FALSE))}

\item{alpha}{Type I error probability}
}
\value{
A data frame object that contains upper and lower band boundaries
}
\description{
Creates Functional Bootstraped (statistical) Bands.
IMPORTANT NOTE: Currently, the script is designed for balanced data sets.
Unbalanced designs (unequal number of curves) may lead to errors!
}
\examples{
load("curvesample.RData")
band.limits <- band(data = curves, type = "prediction", B = 1000, iid = TRUE)
}
\keyword{datasets}
