% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Example data set containing time series (aka curve) data}
\format{
A data frame consisting of 11 rows and 110 columns. Each row
represents a subject, each column a curve point.
}
\usage{
data(data)
}
\description{
The data set consists of 110 curves (i.e, time series). The data set has a
hierarchical structure: It contains 10 curves each from 11  subjects.
All curves are of equal length (n = 101 curve points).
}
\keyword{dataset}
